+++ src/main/java/com/ttasjwi/cafe/controller/UserController.java
@@ -1,0 +1,15 @@
+package com.ttasjwi.cafe.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/users")
+public class UserController {
+
+    @GetMapping("/new")
+    public String createForm() {
+        return "/users/createUserForm";
+    }
+}