+++ src/test/java/com/ttasjwi/cafe/controller/UserControllerTest.java
@@ -1,0 +1,33 @@
+package com.ttasjwi.cafe.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+@WebMvcTest(controllers = UserController.class)
+@ExtendWith(SpringExtension.class)
+class UserControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    @DisplayName("\"/users/new\"로 요청하면 \"/users/createUserForm\" 이 반환된다.")
+    void createFormTest() throws Exception {
+        String requestUrl = "/users/new";
+
+        mockMvc.perform(get(requestUrl))
+                .andExpect(status().isOk())
+                .andExpect(view().name("/users/createUserForm"));
+    }
+}